Revert "refactor: Convert test runner to TypeScript (#1701)"
This reverts commit f30e619fae5deebcbc9340ba834ea06235de994a.
